+++ src/benchmark.rs
@@ -157,6 +157,9 @@ macro_rules! benchmark_config {
         /// Panics if the number of resamples is set to zero
         pub fn nresamples(mut self, n: usize) -> Self {
             assert!(n > 0);
+            if n <= 1000 {
+                println!("\nWarning: It is not recommended to reduce nresamples below 1000.");
+            }
 
             self.config.nresamples = Some(n);
             self