+++ src/lib.rs
@@ -834,6 +834,9 @@ impl<M: Measurement> Criterion<M> {
     /// Panics if the number of resamples is set to zero
     pub fn nresamples(mut self, n: usize) -> Criterion<M> {
         assert!(n > 0);
+        if n <= 1000 {
+            println!("\nWarning: It is not recommended to reduce nresamples below 1000.");
+        }
 
         self.config.nresamples = n;
         self